Add closeProcess method to DlManager
This function ensures that the callback for when a video is done is
executed at the end of the download, not the beginning, and prints that
the video is done downloading to log.

The correct ordering of the video done callback means that errors
prevent video ids from being written to the dictionary.